Working Update
Nothing actually changed here, this is just a notice that everything
should be working just fine now. The only major bug I have encountered
is that the .htaccess file which manages the RewriteModule doesn't work
with the configurable URL code length. It currently only works with 6
character codes using the charset "a-zA-Z0-9", which is fine for now,
but that will probably be in the next major update.